refactored function now better
