+++ app/src/main/java/com/example/beer_mat/tabs/food/FoodScreen.kt
@@ -1,7 +1,6 @@
 package com.example.beer_mat.tabs.food
 
 
-import android.app.AlertDialog
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.runtime.Composable
@@ -11,20 +10,19 @@ import androidx.compose.material3.Text
 import androidx.compose.runtime.*
 import androidx.compose.ui.platform.LocalContext
 import com.example.beer_mat.components.AddFloatingActionButton
+import com.example.beer_mat.components.ShowDialog
 
 class FoodScreen {
     @Composable
     fun Content(modifier: Modifier = Modifier) {
-        val context = LocalContext.current
         var showDialog by remember { mutableStateOf(false) }
 
-        if (showDialog) {
-            val builder = AlertDialog.Builder(context)
-            builder.setTitle("Food").setMessage("Food added")
-            val dialog = builder.create()
-            dialog.show()
-            showDialog = false
-        }
+        ShowDialog(
+            showDialog = showDialog,
+            title = "Food",
+            message = "Food added",
+            onDismiss = { showDialog = false }
+        )
 
         Box(modifier = modifier.fillMaxSize()) {
             Text(text = "Food Content", modifier = Modifier.align(Alignment.TopStart))
